+++ angular-workspace/projects/example-client-app/src/app/customapp/customapp.component.ts
@@ -1,6 +1,11 @@
 /* eslint-disable no-alert */
-import { Component } from '@angular/core';
-import { Combobox, DrawerLocation, MenuItem } from '@ni/nimble-angular';
+import { Component, ViewChild } from '@angular/core';
+import { DrawerLocation, MenuItem, NimbleDialogDirective, OptionNotFound, OPTION_NOT_FOUND, USER_DISMISSED } from '@ni/nimble-angular';
+
+interface ComboboxItem {
+    first: string;
+    last: string;
+}
 
 @Component({
     selector: 'example-customapp',
@@ -8,17 +13,43 @@ import { Combobox, DrawerLocation, MenuItem } from '@ni/nimble-angular';
     styleUrls: ['./customapp.component.scss']
 })
 export class CustomAppComponent {
+    @ViewChild('dialog', { read: NimbleDialogDirective }) public dialog: NimbleDialogDirective<string>;
+    public dialogCloseReason: string;
     public drawerLocation: DrawerLocation = DrawerLocation.right;
     public isDrawerPinned = false;
     public drawerLocations = DrawerLocation;
-    public comboboxValue: string;
+    public comboboxItems: ComboboxItem[] = [
+        { first: 'foo', last: 'bar' },
+        { first: 'Bubba', last: 'Hotep' },
+        { first: 'Mister', last: 'Smithers' }
+    ];
+
+    public comboboxSelectedOption?: ComboboxItem;
+    public comboboxSelectedLastName = this.comboboxSelectedOption?.last;
 
     public onMenuButtonMenuChange(event: Event): void {
         const menuItemText = (event.target as MenuItem).innerText;
         alert(`${menuItemText} selected`);
     }
 
-    public onComboboxChange(event: Event): void {
-        this.comboboxValue = (event.target as Combobox).value;
+    public onDrawerLocationChanged(value: DrawerLocation): void {
+        alert(`drawerLocation: ${value}`);
+    }
+
+    public onComboboxChange(value: ComboboxItem | OptionNotFound): void {
+        if (value !== OPTION_NOT_FOUND) {
+            this.comboboxSelectedLastName = value.last;
+        } else {
+            this.comboboxSelectedLastName = 'not found';
+        }
+    }
+
+    public async openDialog(): Promise<void> {
+        const closeReason = await this.dialog.show();
+        this.dialogCloseReason = (closeReason === USER_DISMISSED) ? 'escape pressed' : closeReason;
+    }
+
+    public closeDialog(reason: string): void {
+        this.dialog.close(reason);
     }
 }