+++ packages/nimble-components/src/combobox/styles.ts
@@ -4,12 +4,16 @@ import {
     failColor,
     bodyDisabledFontColor,
     borderRgbPartialColor,
-    smallPadding
+    smallPadding,
+    borderHoverColor,
+    borderWidth
 } from '../theme-provider/design-tokens';
 
 import { styles as dropdownStyles } from '../patterns/dropdown/styles';
 import { styles as errorStyles } from '../patterns/error/styles';
 import { focusVisible } from '../utilities/style/focus';
+import { appearanceBehavior } from '../utilities/style/appearance';
+import { DropdownAppearance } from '../select/types';
 
 export const styles = css`
     ${dropdownStyles}
@@ -18,6 +22,9 @@ export const styles = css`
     :host {
         --ni-private-hover-bottom-border-width: 2px;
         --ni-private-bottom-border-width: 1px;
+        --ni-private-height-within-border: calc(
+            ${controlHeight} - 2 * ${borderWidth}
+        );
     }
 
     :host([disabled]) *,
@@ -26,8 +33,16 @@ export const styles = css`
         color: ${bodyDisabledFontColor};
     }
 
+    :host(.invalid)::after {
+        border-bottom-color: ${failColor};
+    }
+
     .control {
         bottom-border-width: var(--ni-private-bottom-border-width);
+    }
+
+    .control:focus-within {
+        border-bottom-color: ${borderHoverColor};
     }
 
     :host(.invalid) .control {
@@ -38,8 +53,9 @@ export const styles = css`
         border-color: rgba(${borderRgbPartialColor}, 0.1);
     }
 
-    :host(.invalid[disabled]) .control {
-        border-color: ${failColor};
+    :host(.invalid[disabled]) .control,
+    :host(.open.invalid) .control {
+        border-bottom-color: ${failColor};
     }
 
     .selected-value {
@@ -49,8 +65,8 @@ export const styles = css`
         color: inherit;
         margin: auto 0;
         width: 100%;
-        font-size: inherit;
-        padding-left: 0px;
+        font: inherit;
+        height: var(--ni-private-height-within-border);
     }
 
     .selected-value:hover,
@@ -91,4 +107,14 @@ export const styles = css`
     :host(:empty) .listbox {
         display: none;
     }
-`;
+`.withBehaviors(
+        appearanceBehavior(
+            DropdownAppearance.block,
+            css`
+            :host(.invalid) .control {
+                border-bottom-width: ${borderWidth};
+                padding-bottom: 0;
+            }
+        `
+        )
+    );