+++ packages/nimble-components/src/select/types.ts
@@ -1,0 +1,6 @@
+/**
+ * Types of dropdown appearance.
+ * @public
+ */
+
+export { DropdownAppearance } from '../patterns/dropdown/types';