+++ packages/nimble-components/src/theme-provider/design-token-names.ts
@@ -23,6 +23,7 @@ export const tokenNames: { readonly [key in TokenName]: string } = {
     failColor: 'fail-color',
     warningColor: 'warning-color',
     passColor: 'pass-color',
+    informationColor: 'information-color',
     borderHoverColor: 'border-hover-color',
     iconColor: 'icon-color',
     popupBoxShadowColor: 'popup-box-shadow-color',
@@ -179,3 +180,27 @@ export const scssInternalPropertySetterMarkdown = (
     tokenName: string,
     property: string
 ): string => `\`#{$${prefix}-internal-${tokenName}}: ${property};\``;
+
+// Order of suffixes in the array matter, as we want single word suffixes after the multi-word ones
+const tokenSuffixes = [
+    'RgbPartialColor',
+    'FontColor',
+    'FontLineHeight',
+    'FontWeight',
+    'FontSize',
+    'TextTransform',
+    'FontFamily',
+    'Font',
+    'Size',
+    'Width',
+    'Height',
+    'Delay',
+    'Padding',
+    'Color'
+] as const;
+export type TokenSuffix = typeof tokenSuffixes[number];
+export const suffixFromTokenName = (
+    tokenName: string
+): TokenSuffix | undefined => tokenSuffixes[
+    tokenSuffixes.findIndex(tokenSuffix => tokenName.endsWith(tokenSuffix))
+];