+++ packages/nimble-blazor/CHANGELOG.md
@@ -1,8 +1,56 @@
 # Change Log - @ni/nimble-blazor
 
-This log was last generated on Mon, 20 Jun 2022 20:39:36 GMT and should not be manually modified.
+This log was last generated on Wed, 10 Aug 2022 23:35:50 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 5.5.0
+
+Wed, 10 Aug 2022 23:35:50 GMT
+
+### Minor changes
+
+- Nimble Tooltip Blazor Integration (#309) ([ni/nimble@ab33209](https://github.com/ni/nimble/commit/ab3320969c08460b6449bffa74e76a07d4d3277a))
+
+## 5.4.0
+
+Wed, 10 Aug 2022 21:38:38 GMT
+
+### Minor changes
+
+- Add Appearance to Combobox. ([ni/nimble@32b4ed9](https://github.com/ni/nimble/commit/32b4ed97aaaea52ff211118a405fd34922a3c0cd))
+
+## 5.3.5
+
+Tue, 09 Aug 2022 19:30:28 GMT
+
+### Patches
+
+- Adding 'Placeholder' APIs to Combobox. ([ni/nimble@2b7f674](https://github.com/ni/nimble/commit/2b7f6744f4c9f1504b3852aa059f2b1f1676222c))
+
+## 5.3.0
+
+Mon, 01 Aug 2022 17:08:37 GMT
+
+### Minor changes
+
+- Add Blazor support for nimble-card-button ([ni/nimble@dd8ef82](https://github.com/ni/nimble/commit/dd8ef820cdbe34a6091ce0140228ae0bdbe6f8cb))
+
+## 5.2.0
+
+Wed, 20 Jul 2022 18:41:04 GMT
+
+### Minor changes
+
+- Add appearance to select wrapper and example project ([ni/nimble@a188d7b](https://github.com/ni/nimble/commit/a188d7baa54e1061ad07ffa65b0e5d0a2b495238))
+
+## 5.1.7
+
+Mon, 18 Jul 2022 18:28:07 GMT
+
+### Patches
+
+- Make NI.CSharp.Analyzers a private dependency ([ni/nimble@a88aa03](https://github.com/ni/nimble/commit/a88aa03c588aeb0b2eeccf2f81b897aaf715d3ca))
 
 ## 5.1.0